% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.R
\name{get_A_opt}
\alias{get_A_opt}
\title{Get optimal A matrix from the dual norm representation}
\usage{
get_A_opt(B, C, mu)
}
\arguments{
\item{B}{The J x K regression coefficient estimates.}

\item{C}{The L1 penalization matrix, returned by \code{get_C()}.}

\item{mu}{The smoothing parameter.}
}
\value{
The optimal A matrix defined by Lemma 1.
}
\description{
Get optimal A matrix from the dual norm representation
}
\references{
Smoothing proximal gradient method for general structured sparse regression
}

