% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accgrad.R
\name{get_B_next}
\alias{get_B_next}
\title{Get next value of B in gradient descent}
\usage{
get_B_next(W, grad_f, L, lambda)
}
\arguments{
\item{W}{The W matrix in Algorithm 1 of the reference}

\item{grad_f}{The gradient computed in Algorithm 1}

\item{L}{The lipshitz constant, which determines the step size}

\item{lambda}{The l1 regularization parameter.}
}
\description{
Get next value of B in gradient descent
}

