% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.R
\name{get_C}
\alias{get_C}
\title{Get C matrix giving L1 penalization}
\usage{
get_C(H, lambda, gamma)
}
\arguments{
\item{H}{The H matrix output by \code{get_H()}.}

\item{lambda}{The l1 regularization parameter.}

\item{gamma}{The graph regularization parameter.}
}
\value{
THe C matrix defined on page 6 of the reference.
}
\description{
Get C matrix giving L1 penalization
}
\references{
Graph-Structured Multi-task Regression adn Efficient Optimization for General Fused Lasso
}

