% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proxgrad.R
\name{get_grad_f}
\alias{get_grad_f}
\title{Calculate the gradient for one step}
\usage{
get_grad_f(X, Y, B, C, mu)
}
\arguments{
\item{X}{The data matrix.}

\item{Y}{The matrix of regression responses.}

\item{B}{The J x K regression coefficient estimates.}

\item{C}{The L1 penalization matrix, returned by \code{get_C()}.}

\item{mu}{The smoothing parameter.}
}
\value{
The gradient descent direction for B, given the current estimates.
See equation (11) in the reference.
}
\description{
Calculate the gradient for one step
}
\references{
Graph-Structured Multi-task Regression adn Efficient Optimization for General Fused Lasso
}

