% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.R
\name{get_Lu}
\alias{get_Lu}
\title{Get automatic step-sizes}
\usage{
get_Lu(X, H, lambda, gamma, mu)
}
\arguments{
\item{X}{The data matrix.}

\item{H}{The matrix H defined in the reference.}

\item{lambda}{The l1 regularization parameter.}

\item{gamma}{The graph regularization parameter.}

\item{mu}{The smoothing parameter.}
}
\value{
Lu The automatically chosen step sizes defined in the reference.
}
\description{
Get automatic step-sizes
}
\references{
Smoothing Proximal Gradient Method for General Structured Sparse Regressoin
}

