% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gflasso.R
\name{gflasso}
\alias{gflasso}
\title{Graph-guided fused Lasso via Smoothed Proximal Gradient Descent}
\usage{
gflasso(Y, X, R, opts = list())
}
\arguments{
\item{Y}{The matrix of regression responses, scaled and centered as necessary.}

\item{X}{The data matrix, scaled and centered as necessary.}

\item{R}{The matrix of (thresholded) correlations between columns of Y}

\item{opts}{A potentially partially specified list of regularization and
gradient descent parameters. See merge_proxgrad_opts().}
}
\value{
A list containing the following quantities: \cr
  $B The estimated beta coefficient matrix. Its rows correspond columns of
   X, while its columns correspond to columns of Y. \cr
  $obj The graph fused lasso objective function, over iterations.
  $Lu The automatically calculated step size. \cr
reference.
}
\references{
Smoothing Proximal Gradient Method for General Structured Sparse Regression
}
