% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accgrad.R
\name{objective}
\alias{objective}
\title{Get the objective of the current estimates}
\usage{
objective(X, B, Y, C, lambda)
}
\arguments{
\item{X}{The data matrix.}

\item{B}{The J x K regression coefficient estimates.}

\item{C}{The L1 penalization matrix, returned by \code{get_C()}.}

\item{R}{The matrix of (thresholded) correlations between columns of Y}
}
\value{
The error of the reconstruction according to the original (nonsmooth)
objective.
}
\description{
Get the objective of the current estimates
}

