% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gflasso.R
\name{predict_gflasso}
\alias{predict_gflasso}
\title{GFLASSO prediction}
\usage{
predict_gflasso(model, new.data)
}
\arguments{
\item{model}{The model object from `gflasso`.}

\item{new.data}{The data from which to predict.}
}
\value{
A n x k matrix carrying all k predicted responses across all n samples
}
