% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proxgrad.R
\name{proxgrad}
\alias{proxgrad}
\title{Proximal Gradient Descent for Graph Fused Lasso}
\usage{
proxgrad(Y, X, R, opts = list())
}
\arguments{
\item{Y}{The matrix of regression responses.}

\item{X}{The data matrix.}

\item{R}{The matrix of (thresholded) correlations between columns of Y}

\item{opts}{A potentially partially specified list of regularization and
gradient descent parameters. See merge_proxgrad_opts().}
}
\value{
A list containing the estimated parameters B, as well as various
intermediate gradient descent quantities described in Algorithm 1 of the
reference.
}
\description{
Proximal Gradient Descent for Graph Fused Lasso
}
\references{
Smoothing proximal gradient method for general structured sparse regression
}

