% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accgrad.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Standard soft-thresholding operator}
\usage{
soft_threshold(v, lambda)
}
\arguments{
\item{v}{The vector to soft threshold}

\item{lambda}{The amount to soft-threshold by}
}
\description{
Standard soft-thresholding operator
}

