% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proxgrad.R
\name{update_Z}
\alias{update_Z}
\title{Get the Z update in one step of the gradient descent}
\usage{
update_Z(Lu, grad_f, t)
}
\arguments{
\item{Lu}{The automatically chosen step sizes defined by equation (12)}

\item{grad_f}{A list of gradient descent directions for B at each iteration.}

\item{t}{The current iteration of the gradient descent.}
}
\description{
Get the Z update in one step of the gradient descent
}

