% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_alpha_opt}
\alias{get_alpha_opt}
\title{Get optimal alpha matrix from the dual norm representation}
\usage{
get_alpha_opt(C, W, mu)
}
\arguments{
\item{C}{The L1 penalization matrix, returned by \code{get_C()}.}

\item{mu}{The smoothing parameter.}

\item{B}{The J x K regression coefficient estimates.}
}
\value{
The optimal A matrix defined by Lemma 1.
}
\references{
Smoothing Proximal Gradient Method for General Structured Sparse Regression
}

