% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accgrad.R
\name{get_alpha_opt}
\alias{get_alpha_opt}
\title{Get Optimal Alpha
S(C * W' / mu) where S projects into the interval [-1, 1].}
\usage{
get_alpha_opt(C, W, mu)
}
\arguments{
\item{C}{The L1 penalization matrix, returned by \code{get_C()}.}

\item{W}{The W matrix in Algorithm 1 of the reference}

\item{mu}{The smoothing parameter.}
}
\value{
S(C * W' / mu)
}
\description{
Get Optimal Alpha
S(C * W' / mu) where S projects into the interval [-1, 1].
}

