% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accgrad.R
\name{get_grad_f}
\alias{get_grad_f}
\title{Calculate the gradient for one step}
\usage{
get_grad_f(X, Y, W, C, mu)
}
\arguments{
\item{X}{The data matrix.}

\item{Y}{The matrix of regression responses.}

\item{C}{The L1 penalization matrix, returned by \code{get_C()}.}

\item{mu}{The smoothing parameter.}

\item{B}{The J x K regression coefficient estimates.}
}
\value{
The gradient descent direction for B, given the current estimates.
See equation (11) in the reference.
}
\references{
Smoothing Proximal Gradient Method for General Structured Sparse Regressoin
}

