% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.R
\name{get_H}
\alias{get_H}
\title{Get H matrix giving Graph Penalization}
\usage{
get_H(R)
}
\arguments{
\item{R}{The matrix of (thresholded) correlations between columns of Y}
}
\value{
The matrix H defined in the reference.
}
\description{
Get H matrix giving Graph Penalization
}
\references{
Smoothing Proximal Gradient Method for General Structured Sparse Regressoin
}

