% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.R
\name{get_Lu}
\alias{get_Lu}
\title{Get automatic step-sizes}
\usage{
get_Lu(X, R, lambda, gamma, mu)
}
\arguments{
\item{X}{The data matrix.}

\item{R}{The matrix of (thresholded) correlations between columns of Y}

\item{lambda}{The l1 regularization parameter.}

\item{gamma}{The graph regularization parameter.}

\item{mu}{The smoothing parameter.}
}
\value{
Lu The automatically chosen step sizes defined by equation (12)
}
\description{
Get automatic step-sizes
}
\references{
Graph-Structured Multi-task Regression adn Efficient Optimization for General Fused Lasso
}

