% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gflasso.R
\name{merge_proxgrad_opts}
\alias{merge_proxgrad_opts}
\title{Merge default proximal gradient descent options}
\usage{
merge_proxgrad_opts(opts, J, K)
}
\arguments{
\item{opts}{A potentially partially specified list of regularization and
gradient descent parameters. The currently supported options are,
$delta_conv How small does the change in B need to be to declare convergence?
$eps A tolerance used in calculating the degree of smoothing in the proximal
gradient objective.
$gamma The graph regularization parameter.
$iter_max What is the maximum number of iterations we should run?
$lambda The l1 regularization parameter.
$verbose Should the gradient descent print its progress?}
}
\value{
A modified version of opts with defaults filled in.
}

